Keeping the manual PowerShell in line with MSI
The MSI installs bicep to the "%LOCALAPPDATA%\Programs\Bicep CLI" folder, so updating the script to keep consistency between deployments. Also, removed the attribute to hide the folder as that is not common to do with deploy folders, could make it harder for users, and lends itself to appear nefarious.